+++ jwt-util/jwt_util/go/core/conversions.go
@@ -1,0 +1,1 @@
+package core
